+++ src/grimp/__init__.py
@@ -2,7 +2,7 @@ __version__ = "3.1"
 
 from .application.ports.graph import DetailedImport, ImportGraph
 from .domain.analysis import PackageDependency, Route
-from .domain.valueobjects import DirectImport, Module
+from .domain.valueobjects import DirectImport, Module, Level
 from .main import build_graph
 
 __all__ = [
@@ -13,4 +13,5 @@ __all__ = [
     "PackageDependency",
     "Route",
     "build_graph",
+    "Level",
 ]